Added TypeError to delete_event, Created tests for delete_event, Formally documented Test Strategy.md
Added TypeError to delete_event
now a boolean is made true if the event deletes, and is parsed back for the means of testing.

Created ests for delete_event
test with 'bad event' does not work yet

Formally documented Test Strategy.md
begun transfering test strategy to the required .md file. Link to LucidChart with program CFGs included in file